Update README example to use 0.4.0
As a new user (to typst and to modern-cv), I copy-pasted the example, got it working, and then started modifying using features released post 0.1.0 (such as `colored-headers`). It took me a while to figure out that the import is of a very old version.

PS: thank you for this package, works very well.